+++ src/firestore/write/setIn.ts
@@ -1,7 +1,9 @@
-import { useFire } from '../context';
-import { timestamp } from '../utils/addTimestamp';
-import { isEmpty } from '../utils/isEmpty';
-import { ReturnWithDoc } from './SetReturn';
+import { useState } from 'react';
+import { useFire } from '../../context';
+import { timestamp } from '../../utils/addTimestamp';
+import { isEmpty } from '../../utils/isEmpty';
+
+type setFunction = (data: object, doc: string) => [boolean, null | Error]
 
 /**
  * Hook for setting the data. Takes an object with these params:
@@ -10,11 +12,13 @@ import { ReturnWithDoc } from './SetReturn';
  * @param callback - optional function to be called back after success
  * @returns a function that returns returns a promise with no return value after resolution
  */
-export const useSetFS = (
+export const setIn = (
   collection: string,
-  options: {merge: boolean} | undefined,
-): ReturnWithDoc => {
+  merge: boolean = false,
+): setFunction => {
   const { firestore } = useFire();
+  const [loading, setLoading] = useState<boolean>(false);
+  const [error, setError] = useState<null | Error>(null);
 
   /**
    * Function used to set a document in the firestore
@@ -24,6 +28,7 @@ export const useSetFS = (
    */
   // this `| undefined` is really killing me but i don't know how to deal with it
   const setFunction = (data: object, doc: string | undefined) => {
+    setLoading(true);
     if (isEmpty(data)) throw new Error('You need to specify the data to update to');
     if (!doc) throw new Error('You need to pass an uid of the document you want to update');
