+++ src/firestore/write/useUpdateFS.ts
@@ -1,15 +1,12 @@
-import { useFire } from '../context';
-import { timestamp } from '../utils/addTimestamp';
-import { ReturnWithDoc } from './SetReturn';
-import { isEmpty } from '../utils/isEmpty';
+import { useFire } from '../../context';
+import { timestamp } from '../../utils/addTimestamp';
+import { ReturnWithDoc } from '../SetReturn';
+import { isEmpty } from '../../utils/isEmpty';
 
 /**
- * Hook for updating the data. Takes an object with these params:
+ * Hook for updating the data
  * @param collection - string pointing to a collection
- * @param doc - string pointing to a document to edit
- * @param data - an object with a field to update in the firestore
- * @param callback - optional function to be called back after success
- * @returns array with `loading` state, `error` object
+ * @returns a promise that resolves to void and will update a document in firestore
  */
 export const useUpdateFS = (collection: string): ReturnWithDoc => {
   const { firestore } = useFire();
@@ -18,7 +15,7 @@ export const useUpdateFS = (collection: string): ReturnWithDoc => {
    * Function used to set a document in the firestore
    * @param data - an object that will be set in the firestore
    * @param doc - string pointing to a document to edit
-   * @returns a promise that resolves to void and will set an object in firestore
+   * @returns a promise that resolves to void and updates a document in firestore
    */
   // this `| undefined` is really killing me but i don't know how to deal with it
   const updateFunction = (data: object, doc: string | undefined) => {
